refactor: extract artifact panel from AgentWorkspace
